backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langgraph
pydantic
python-dotenv
sqlite-utils
